% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calCutOff.R
\name{calCutOff}
\alias{calCutOff}
\title{This function is used to calculate the cut off values for low-risk and
high-risk groups. The values are calculated by the proportion between
converters with a value and all patients having this certain value
corresponding with the given pHigh and pLow values. For each feature cutoff
calculation all patients that have not been categorized into the high or low
risk categories will be used.}
\usage{
calCutOff(inputData, pHigh, pLow)
}
\arguments{
\item{inputData}{input data given by the user (type: data.frame)}

\item{pHigh}{risk-threshold, defined by the user (type: double, value between 0 and 1)}

\item{pLow}{risk-threshold, defined by the user (type: double, value between 0 and 1)}
}
\value{
fit, a variable which contains links to the calculated cutOff table, filtered data and effectiveness of the cutOff values
}
\description{
This function is used to calculate the cut off values for low-risk and
high-risk groups. The values are calculated by the proportion between
converters with a value and all patients having this certain value
corresponding with the given pHigh and pLow values. For each feature cutoff
calculation all patients that have not been categorized into the high or low
risk categories will be used.
}
