% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictTree.R
\name{predictTree}
\alias{predictTree}
\title{predictTree
This function predicts the risk of a patient conversion to alzheimers, based
on the user given data/parameters and previously calculated cutoff values (calculated in design tree).}
\usage{
predictTree(cutOff, inputData)
}
\arguments{
\item{cutOff}{the calculated cutOff values for each parameter (type: matrix)}

\item{inputData}{the inputData / patient, that the previously calculated cutOff values are tested on (type: matrix)}
}
\value{
result, variable result with its links to the cutOff matrix and predicted patient information
}
\description{
predictTree
This function predicts the risk of a patient conversion to alzheimers, based
on the user given data/parameters and previously calculated cutoff values (calculated in design tree).
}
