% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictTree.R
\name{predictTreeRec}
\alias{predictTreeRec}
\title{predictTreeRec
This recursive function calculates the cutoff value for each parameter
dependent upon the patient ratio, it also verifies if the model is effective
(cut_high > cut_low).}
\usage{
predictTreeRec(cutOff, inputData)
}
\arguments{
\item{cutOff}{the calculated cutOff values for each parameter (type: matrix)}

\item{inputData}{the inputData / patient, that the previously calculated cutOff values are tested on (type: matrix)}
}
\value{
result, variable result with its links to the cutOff matrix and predicted patient information
}
\description{
predictTreeRec
This recursive function calculates the cutoff value for each parameter
dependent upon the patient ratio, it also verifies if the model is effective
(cut_high > cut_low).
}
