% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareData.R
\name{prepareData}
\alias{prepareData}
\title{prepareData
prepare Data is a function to transfer the given data into a data frame which contains only the information
needed for the Triple Sec algorithm.}
\usage{
prepareData(data, firstFeature, featureList)
}
\arguments{
\item{data}{the data table with patients and all available data}

\item{firstFeature}{the first Feature which should be set}

\item{featureList}{the other features, excluding the first feature}
}
\value{
inputData, data table including only all relevant information for the Triple Sec algorithm
}
\description{
prepareData
prepare Data is a function to transfer the given data into a data frame which contains only the information
needed for the Triple Sec algorithm.
}
