% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruneTree.R
\name{pruneTree}
\alias{pruneTree}
\title{pruneTree
pruneTree splits the given data table into training and test data. Then calls calCutOff for the training data
and predictTree with the calculated cutOff values of calCutOff and the test data.}
\usage{
pruneTree(inputData, pHigh, pLow)
}
\arguments{
\item{inputData}{input data table, given by the user (shoud be a data frame)}

\item{pHigh}{risk-threshold, defined by the user (should be a percentage value!)}

\item{pLow}{risk-threshold, defined by the user (should be a percentage value!)}
}
\value{
result, a variable with a link to the average cutoff values for each feature and a table data with the patients and feature and their risk assesment
}
\description{
pruneTree
pruneTree splits the given data table into training and test data. Then calls calCutOff for the training data
and predictTree with the calculated cutOff values of calCutOff and the test data.
}
