% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictTree.R
\name{predictTree}
\alias{predictTree}
\title{This function calls the recursive function predictTreeRec. The function
returns a variable with a link to a patient data table containing their risk
assessments and a link to the average cutoff value matrix.}
\usage{
predictTree(cutOff, inputData)
}
\arguments{
\item{cutOff}{the calculated cutOff values for each parameter (type: matrix)}

\item{inputData}{the inputData / patient, that the previously calculated cutOff values are tested on (type: matrix)}
}
\value{
result, variable result with its links to the cutOff matrix and predicted patient information
}
\description{
This function calls the recursive function predictTreeRec. The function
returns a variable with a link to a patient data table containing their risk
assessments and a link to the average cutoff value matrix.
}
