% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareData.R
\name{prepareData}
\alias{prepareData}
\title{prepareData
This function transforms the given data into a data frame containing only the columns necessary for the TripleSec algorithm, therefor for the risk assessment.}
\usage{
prepareData(data, firstFeature, featureList)
}
\arguments{
\item{data}{the data table with patients and all available data}

\item{firstFeature}{the first Feature which should be set}

\item{featureList}{the other features, excluding the first feature}
}
\value{
inputData, data table including only all relevant information for the Triple Sec algorithm
}
\description{
prepareData
This function transforms the given data into a data frame containing only the columns necessary for the TripleSec algorithm, therefor for the risk assessment.
}
