% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruneTree.R
\name{pruneTree}
\alias{pruneTree}
\title{pruneTree
This function divides the given input data frame into training and test data.
It calls the function \code{\link{calCutOff}} for the training data, which
returns a matrix containing the calculated cut off values. Using this matrix
and the test data, pruneTree then calls \code{\link{predictTree}} for the
test data. \code{\link{predictTree}} returns the risk assessment for the test
data to pruneTree.}
\usage{
pruneTree(inputData, pHigh, pLow)
}
\arguments{
\item{inputData}{input data table, given by the user (shoud be a data frame)}

\item{pHigh}{risk-threshold, defined by the user (should be a percentage value!)}

\item{pLow}{risk-threshold, defined by the user (should be a percentage value!)}
}
\value{
result, a variable with a link to the average cutoff values for each feature and a table data with the patients and feature and their risk assessment
}
\description{
pruneTree
This function divides the given input data frame into training and test data.
It calls the function \code{\link{calCutOff}} for the training data, which
returns a matrix containing the calculated cut off values. Using this matrix
and the test data, pruneTree then calls \code{\link{predictTree}} for the
test data. \code{\link{predictTree}} returns the risk assessment for the test
data to pruneTree.
}
